requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.30.0
faiss-cpu>=1.7.4
numpy>=1.26.0
networkx>=3.2.0
